Dedicated trait for sending direct responses
Will be useful later if I ever get around to implementing file responder
