+++ src/Responder/Concerns/ModifiesResponseHtml.php
@@ -8,18 +8,10 @@ namespace ToyWpRouting\Responder\Concerns;
 trait ModifiesResponseHtml
 {
     protected array $modifiesResponseHtmlData = [
-        'body' => null,
         'bodyClasses' => [],
         'title' => null,
         'template' => null,
     ];
-
-    public function withBody(string $body): self
-    {
-        $this->modifiesResponseHtmlData['body'] = $body;
-
-        return $this;
-    }
 
     public function withBodyClass(string $bodyClass): self
     {
@@ -77,28 +69,7 @@ trait ModifiesResponseHtml
             return $this->modifiesResponseHtmlData['template'];
         });
 
-        // @todo Merge all template_include functionality?
-        $this->addFilter('template_include', function ($template) {
-            if (! is_string($this->modifiesResponseHtmlData['body'])) {
-                return $template;
-            }
-
-            echo $this->modifiesResponseHtmlData['body'];
-
-            return dirname(__DIR__, 3) . '/templates/blank.php';
-        });
-
-        $this->addConflictCheck(function () {
-            // template and json? redirect?
-            // body and json? redirect?
-
-            if (
-                is_string($this->modifiesResponseHtmlData['body'])
-                && is_string($this->modifiesResponseHtmlData['template'])
-            ) {
-                return 'Cannot set both response body and template';
-            }
-        });
+        // @todo All conflict with json? redirect?
     }
 
     protected function isModifyingResponseHtmlTemplate(): bool
